packages/automation-cli/pyproject.toml
[project]
name = "automation-mcp"
version = "1.0.0"
description = "Unified MCP server for Reddit and SEO operations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp~=1.6.0",
    "fastmcp~=2.2.0",
    "pydantic~=2.10.0",
    "redditwarp>=1.0.0",
    "requests~=2.32.0",
    "playwright~=1.51.0",
    "praw~=7.8.0",
    "python-dotenv~=1.0.0",
    "google-api-python-client~=2.150.0",
    "google-auth~=2.35.0",
    "google-auth-oauthlib~=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/automation_mcp"]

[project.scripts]
automation-mcp = "automation_mcp.server:main"
automation-cli = "automation_mcp.cli:main"
